Fixed a minor bug in checkpointing EEXE simulations and allow increasing the number of iterations in EEXE

+++ ensemble_md/cli/run_EEXE.py
@@ -110,22 +110,29 @@ def main():
         start_idx = 1
     else:
         if rank == 0:
-            # If there is a checkpoint file, we see the execution as an exntension of an EEXE simulation
+            # If there is a checkpoint file, we see the execution as an extension of an EEXE simulation
             ckpt_data = np.load(args.ckpt)
-            start_idx = len(ckpt_data[0]) - 1
+            start_idx = len(ckpt_data[0])
             print(f'\nGetting prepared to extend the EEXE simulation from iteration {start_idx} ...')
 
-            print('Deleting data generated after the checkpoint ...')
+            print('Deleting corrupted data ...')
             corrupted = glob.glob('gmxapi.commandline.cli*')  # corrupted iteration
             corrupted.extend(glob.glob('mdrun*'))
             for i in corrupted:
                 shutil.rmtree(i)
+            if len(corrupted) == 0:
+                corrupt_bool = False
 
             for i in range(EEXE.n_sim):
-                n_finished = len(next(os.walk(f'sim_{i}'))[1])  # number of finished iterations
-                for j in range(start_idx, n_finished):
-                    print(f'  Deleting the folder sim_{i}/iteration_{j}')
-                    shutil.rmtree(f'sim_{i}/iteration_{j}')
+                n_finished = len(next(os.walk(f'sim_{i}'))[1])  # number of finished iterations (the last might be initialized but corrupted though)  # noqa: E501
+                if n_finished == EEXE.n_iter and corrupt_bool is False:
+                    print('Extension aborted: The expected number of iterations have been completed!')
+                    sys.exit()
+                else:
+                    print('Deleting data generated after the checkpoint ...')
+                    for j in range(start_idx, n_finished):
+                        print(f'  Deleting the folder sim_{i}/iteration_{j}')
+                        shutil.rmtree(f'sim_{i}/iteration_{j}')
 
             # Read g_vecs.npy and rep_trajs.npy so that new data can be appended, if any.
             EEXE.rep_trajs = [list(i) for i in ckpt_data]
